Author service,dto,controllers,along with other apis

+++ src/main/java/com/example/demo/models/Book.java
@@ -16,10 +16,11 @@ public class Book {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private Integer id;
 
     private String name;
-    @ManyToOne
+
+    @ManyToOne  // many books can be written by one author, or many rows in the book table can have the same author_id
     @JoinColumn
     private Author author;
 
@@ -31,7 +32,7 @@ public class Book {
     @Enumerated(value = EnumType.STRING)
     private Genre genre;
 
-    @OneToMany(mappedBy = "book")
+    @OneToMany(mappedBy = "book")  // mappedBy tells the hibernate that the book field in the Transaction class is the owner of the relationship this is a bidirectional relationship
     private List<Transaction> transactions;
